:recycle: refactor: 삼항 연산자 제거 및 코드 단순화
