+++ src/components/base/Icon/Icon.tsx
@@ -1,23 +1,26 @@
-import { ReactChild } from 'react';
-import { Send, Reply, Delete } from './Icons';
+import { Send, Reply, Delete } from './IconConverter';
 import styled, { css } from 'styled-components';
 
 interface IconProps {
   name: string;
   size: number;
-  onClick: () => void;
+  onClick?: () => void;
+}
+interface Icons {
+  [key: string]: ({ size }: IconProps) => JSX.Element;
 }
 
+const icons: Icons = {
+  send: Send,
+  reply: Reply,
+  delete: Delete,
+};
+
 const Icon = ({ name, size, onClick }: IconProps) => {
+  const SpecificIcon = icons[name];
   return (
     <StyledIcon name={name} size={size} onClick={onClick}>
-      {name === 'send' ? (
-        <Send size={size} />
-      ) : name === 'reply' ? (
-        <Reply size={size} />
-      ) : (
-        <Delete size={size} />
-      )}
+      <SpecificIcon name={name} size={size} />
     </StyledIcon>
   );
 };
@@ -31,17 +34,22 @@ const StyledIcon = styled.button<IconProps>`
   ${(props) =>
     props.name === 'send'
       ? css`
-          width: ${props.size}px;
-          height: ${props.size}px;
+          width: ${props.size - 1}px;
+          height: ${props.size - 1}px;
+          &:active {
+            background-color: rgba(0, 0, 0, 0.4);
+          }
         `
       : css`
-          width: ${props.size + 4}px;
-          height: ${props.size + 4}px;
+          width: ${props.size + 8}px;
+          height: ${props.size + 8}px;
+          &:hover {
+            box-shadow: 1px 2px 2px rgba(0, 0, 0, 0.2);
+          }
+          &:active {
+            background-color: rgba(0, 0, 0, 0.1);
+          }
         `}
-
-  &:active {
-    background-color: rgba(0, 0, 0, 0.2);
-  }
 `;
 
 export default Icon;
